+++ mayan/apps/document_states/models/workflow_transition_models.py
@@ -1,29 +1,21 @@
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-from mayan.apps.common.validators import YAMLValidator
 from mayan.apps.databases.model_mixins import ExtraDataModelMixin
 from mayan.apps.events.decorators import method_event
 from mayan.apps.events.event_managers import (
     EventManagerMethodAfter, EventManagerSave
 )
-from mayan.apps.events.models import StoredEventType
 
 from ..events import event_workflow_template_edited
-from ..literals import FIELD_TYPE_CHOICES, WIDGET_CLASS_CHOICES
 
 from .workflow_models import Workflow
 from .workflow_state_models import WorkflowState
 from .workflow_transition_model_mixins import (
-    WorkflowTransitionBusinessLogicMixin,
-    WorkflowTransitionFieldBusinessLogicMixin,
-    WorkflowTransitionTriggerEventBusinessLogicMixin
+    WorkflowTransitionBusinessLogicMixin
 )
 
-__all__ = (
-    'WorkflowTransition', 'WorkflowTransitionField',
-    'WorkflowTransitionTriggerEvent'
-)
+__all__ = ('WorkflowTransition',)
 
 
 class WorkflowTransition(
@@ -89,140 +81,6 @@ class WorkflowTransition(
             'action_object': 'self',
             'event': event_workflow_template_edited,
             'target': 'workflow'
-        }
-    )
-    def save(self, *args, **kwargs):
-        return super().save(*args, **kwargs)
-
-
-class WorkflowTransitionField(
-    ExtraDataModelMixin, WorkflowTransitionFieldBusinessLogicMixin,
-    models.Model
-):
-    _ordering_fields = ('label', 'name', 'required', 'widget_kwargs')
-
-    transition = models.ForeignKey(
-        on_delete=models.CASCADE, related_name='fields',
-        to=WorkflowTransition, verbose_name=_(message='Transition')
-    )
-    field_type = models.PositiveIntegerField(
-        choices=FIELD_TYPE_CHOICES, verbose_name=_(message='Type')
-    )
-    name = models.CharField(
-        help_text=_(
-            message='The name that will be used to identify this field in other '
-            'parts of the workflow system.'
-        ), max_length=128, verbose_name=_(message='Internal name')
-    )
-    label = models.CharField(
-        help_text=_(
-            message='The field name that will be shown on the user interface.'
-        ), max_length=128, verbose_name=_(message='Label')
-    )
-    help_text = models.TextField(
-        blank=True, help_text=_(
-            message='An optional message that will help users better understand the '
-            'purpose of the field and data to provide.'
-        ), verbose_name=_(message='Help text')
-    )
-    required = models.BooleanField(
-        default=False, help_text=_(
-            message='Whether this fields needs to be filled out or not to proceed.'
-        ), verbose_name=_(message='Required')
-    )
-    widget = models.PositiveIntegerField(
-        blank=True, choices=WIDGET_CLASS_CHOICES, help_text=_(
-            message='An optional class to change the default presentation of the '
-            'field.'
-        ), null=True, verbose_name=_(message='Widget class')
-    )
-    widget_kwargs = models.TextField(
-        blank=True, help_text=_(
-            message='A group of keyword arguments to customize the widget. '
-            'Use YAML format.'
-        ), validators=[
-            YAMLValidator()
-        ], verbose_name=_(message='Widget keyword arguments')
-    )
-
-    class Meta:
-        unique_together = ('transition', 'name')
-        verbose_name = _(message='Workflow transition field')
-        verbose_name_plural = _(message='Workflow transition fields')
-
-    def __str__(self):
-        return self.label
-
-    @method_event(
-        action_object='self',
-        event_manager_class=EventManagerMethodAfter,
-        event=event_workflow_template_edited,
-        target='transition.workflow'
-    )
-    def delete(self, *args, **kwargs):
-        return super().delete(*args, **kwargs)
-
-    @method_event(
-        event_manager_class=EventManagerSave,
-        created={
-            'action_object': 'self',
-            'event': event_workflow_template_edited,
-            'target': 'transition.workflow'
-        },
-        edited={
-            'action_object': 'self',
-            'event': event_workflow_template_edited,
-            'target': 'transition.workflow'
-        }
-    )
-    def save(self, *args, **kwargs):
-        return super().save(*args, **kwargs)
-
-
-class WorkflowTransitionTriggerEvent(
-    ExtraDataModelMixin, WorkflowTransitionTriggerEventBusinessLogicMixin,
-    models.Model
-):
-    transition = models.ForeignKey(
-        on_delete=models.CASCADE, related_name='trigger_events',
-        to=WorkflowTransition, verbose_name=_(message='Transition')
-    )
-    event_type = models.ForeignKey(
-        on_delete=models.CASCADE, to=StoredEventType,
-        verbose_name=_(message='Event type')
-    )
-
-    class Meta:
-        ordering = ('event_type__name',)
-        unique_together = ('transition', 'event_type')
-        verbose_name = _(message='Workflow transition trigger event')
-        verbose_name_plural = _(
-            message='Workflow transitions trigger events'
-        )
-
-    def __str__(self):
-        return str(self.transition)
-
-    @method_event(
-        action_object='self',
-        event_manager_class=EventManagerMethodAfter,
-        event=event_workflow_template_edited,
-        target='transition.workflow'
-    )
-    def delete(self, *args, **kwargs):
-        return super().delete(*args, **kwargs)
-
-    @method_event(
-        event_manager_class=EventManagerSave,
-        created={
-            'action_object': 'self',
-            'event': event_workflow_template_edited,
-            'target': 'transition.workflow'
-        },
-        edited={
-            'action_object': 'self',
-            'event': event_workflow_template_edited,
-            'target': 'transition.workflow'
         }
     )
     def save(self, *args, **kwargs):