+++ HISTORY.rst
@@ -72,6 +72,8 @@
 - Allow the ``JavaScriptCatalogPublic`` sub class to bypass authentication
   and avoid JavaScript errors for non authenticated users.
 - Add thousand separator to list header number details.
+- Split workflow template fields and workflow template triggers into their
+  own modules.
 
 4.6.2 (2024-03-04)
 ==================